feat: 고도화된 Projects 도구 및 프로젝트 완성
주요 기능 추가:
- Wrapper 상세뷰 구현 (진행률 대시보드, 하위 프로젝트 목록, 완료 처리)
- 스마트 완료 시스템 (프로젝트 완료 시 Wrapper 자동 완료 체크)
- 컨텍스트 메뉴 기반 CRUD 작업 (프로젝트/래퍼 추가/삭제)
- 트리뷰 개선 (완료 섹션 자동 접기, 동적 재구성)
- 안전장치 구현 (하위 항목 보호, 상태 검증)

프로젝트 정리:
- requirements.txt 정리 (핵심 패키지만 유지)
- .gitignore 업데이트 (Claude Code 설정 추가)
- README.md 및 CLAUDE.md 작성 (상세 사용법 및 개발 가이드)
- 전체 코드베이스 점검 및 문서화 완료

+++ src/core/components/sidebar.py
@@ -11,7 +11,7 @@ def emoji_icon(emoji: str, size: int = 42) -> QIcon:
     pixmap.fill(Qt.transparent)
 
     painter = QPainter(pixmap)
-    font = QFont("Segoe UI Emoji", size * 0.7)  # Emoji 전용 폰트 사용
+    font = QFont("Segoe UI Emoji", int(size * 0.7))  # Emoji 전용 폰트 사용
     painter.setFont(font)
     painter.drawText(pixmap.rect(), Qt.AlignCenter, emoji)
     painter.end()
